rename MDS036 to max-section-length
Align with max-file-length naming. "Size" was ambiguous (bytes?
tokens?); "length" mirrors the file-level analog and makes it clear
the limit counts lines. Renames:

- package sectionsizelimits -> maxsectionlength
- rule name section-size-limits -> max-section-length
- fixture dir MDS036-section-size-limits -> MDS036-max-section-length
