+++ internal/rules/MDS036-max-section-length/bad/per-level.md
